Fix bug where wallet USDC is not swept for a fresh account. (#214)
Before this change, `DydxTransferSubaccountWorker` was listening to
`selectedSubaccount` from Abacus. `selectedSubaccount` is basically
`subaccounts[subaccountNumber]` but for a fresh account, the Indexer
does not return ANY subaccounts. so the `subaccounts[subaccountNumber]`
lookup always fails.

The fix here is to reference the `subaccountNumber` on the Abacus state
manager itself. This is fine because even in a multi-parent subaccount
world, `subaccountNumber` is intended to be a parent subaccount.